Toggle Outlines on / off in Modal web server
Working well with outlines -- something funky without where it repeats the question, i.e.: "What is the capital of Bulgaria?" --> "Sofia. What is the capital of Bulgaria? ...." (goes on for awhile like this) -- maybe something in the sampling params I changed accidentally -- same args for extract_assistant_response and model hash